backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cloudinary==1.40.0
python-decouple==3.8
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
email-validator==2.1.1
cryptography>=3.4.8
bcrypt>=3.2.0